+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Bump library versions.
+- Improve documentation on custom conformers.
 
 ## [0.3.0] — 2018-09-30
 ### Added